pyproject.toml
[tool.poetry]
name = "pynumaflow"
version = "0.6.0"
description = "Provides the interfaces of writing Python User Defined Functions and Sinks for NumaFlow."
authors = ["NumaFlow Developers"]
readme = "README.md"
license = "Apache-2.0"
maintainers = [
    "Avik Basu <avikbasu93@gmail.com>",
    "Vigith Maurice <vigith@gmail.com>",
]
classifiers = [
    "Topic :: Software Development :: Libraries",
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11"
]
repository = "https://github.com/numaproj/numaflow-python"

[tool.poetry.dependencies]
python = ">=3.9, <3.12"
grpcio = "^1.48.1"
grpcio-tools = "^1.48.1"
google-cloud = "^0.34.0"
google-api-core = "^2.11.0"
protobuf = ">=3.20,<5.0"
aiorun = "^2023.7"
uvloop = "^0.19.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-cov = "^3.0"
black = "^23.1"
grpcio-testing = "^1.48.1"
ruff = "^0.0.264"
pre-commit = "^3.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py39"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.idea
  | _build
  | buck-out
  | build
  | dist
  | tests/.*/setup.py
  | function/proto
  | sink/proto
)/
'''

[tool.ruff]
line-length = 100
src = ["pynumaflow", "tests", "examples"]
select = ["E", "F", "UP"]
target-version = "py39"
show-fixes = true
show-source = true
extend-exclude = [
    "*_pb2.py",
    "*_pb2*.py",
    "*.pyi"
]
